GEODE-9369: Command to copy region entries from a WAN site to another (#6601)
* GEODE-9369: Command to copy region entries from a WAN site to another

The command has been implemented as proposed in
https://cwiki.apache.org/confluence/display/GEODE/Geode+Command+to+replicate+region+data+from+one+site+to+another+connected+via+WAN
with some modifications with respect to the initial proposal.

The command will get the entries of a region
in a WAN site and will put them in batches
that will be sent by a gateway sender to a remote
WAN site.

* GEODE-9369: Remove comments added to geode-gfsh/src/test/resources/expected-pom.xml

* GEODE-9369: Changes after review

* GEODE-9369: Update with Kirk's review comments

* GEODE-9369: Update with boglesby's review comments

* GEODE-9369: Changes after boglesby's review

* GEODE-9369: Updated with davebarnes97's review comments.

* GEODE-9369: More changes after davebarnes97's review

* GEODE-9369: Updated with DonalEvan's comments

* GEODE-9369: Small refactoring: Use CompletableFuture instead of Callable + FutureTask

* GEODE-9369: Fix race condition added in previous commit

* GEODE-9369: Small change of re-review from DonalEvans

* GEODE-9369: Fix esporadic test failures due to error log.

* GEODE-9369: Changes required after rebasing to develop

+++ geode-core/src/main/java/org/apache/geode/internal/cache/tier/sockets/CacheServerStats.java
@@ -992,8 +992,8 @@ public class CacheServerStats implements MessageStats {
     return this.stats.getDouble(loadPerConnectionId);
   }
 
-  public int getProcessBatchRequests() {
-    return this.stats.getInt(processBatchRequestsId);
+  public long getProcessBatchRequests() {
+    return this.stats.getLong(processBatchRequestsId);
   }
 
   public void close() {
